requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openpyxl==3.1.2
requests==2.31.0
Pillow==10.1.0
gunicorn==21.2.0
flask-login==0.6.3
bcrypt==4.1.2
